Refactor componentes: actualizar nombres y rutas en varios archivos para mejorar la organización del código.
